refactor: 참여자 공통 로직 분리
